added models

+++ model/url.go
@@ -1,7 +1,7 @@
 package model
 
 type URL struct {
-	WebsiteId  int
-	ResourceId int
+	WebsiteId  int64
+	ResourceId int64
 	URL        string
 }
